fix: applied comments
