+++ packages/wallets/src/features/cfd/modals/CTraderSuccessModal/components/CTraderSuccessModalButtons.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import { useHistory } from 'react-router-dom';
+import { WalletButton, WalletButtonGroup } from '../../../../../components';
+
+type TCTraderSuccessModalButtons = {
+    hide: () => void;
+    isDemo: boolean;
+};
+
+const CTraderSuccessModalButtons = ({ hide, isDemo }: TCTraderSuccessModalButtons) => {
+    const history = useHistory();
+
+    return isDemo ? (
+        <div className='wallets-success-btn'>
+            <WalletButton isFullWidth onClick={hide} size='lg'>
+                OK
+            </WalletButton>
+        </div>
+    ) : (
+        <WalletButtonGroup isFlex isFullWidth>
+            <WalletButton onClick={hide} size='lg' variant='outlined'>
+                Maybe later
+            </WalletButton>
+            <WalletButton
+                onClick={() => {
+                    hide();
+                    history.push('/wallets/cashier/transfer');
+                }}
+                size='lg'
+            >
+                Transfer funds
+            </WalletButton>
+        </WalletButtonGroup>
+    );
+};
+
+export default CTraderSuccessModalButtons;
