+++ packages/wallets/src/features/cfd/modals/CTraderSuccessModal/components/index.ts
@@ -1,0 +1,1 @@
+export { default as CTraderSuccessModalButtons } from './CTraderSuccessModalButtons';
